Gráfico de temperatura promedio mensual por estación

Gráfico de temperatura promedio mensual por estación

Usage

grafico_temperatura_mensual(
  datos,
  colores = "darkgreen",
  titulo = "Temperatura"
)

Arguments

datos

Un data frame que contenga al menos las columnas fecha, id y temperatura_abrigo_150cm.

colores

Un vector con los colores a utilizar para las estaciones. Por defecto "darkgreen".

titulo

Título del gráfico. Por defecto "Temperatura".

Value

Un objeto ggplot con el gráfico generado.

Examples

grafico_temperatura_mensual(datos)
#> Error in grafico_temperatura_mensual(datos): could not find function "grafico_temperatura_mensual"
grafico_temperatura_mensual(datos, colores = c("red", "blue"), titulo = "Temperatura mensual promedio")
#> Error in grafico_temperatura_mensual(datos, colores = c("red", "blue"),     titulo = "Temperatura mensual promedio"): could not find function "grafico_temperatura_mensual"